requirements.txt
requests
beautifulsoup4
pandas
numpy
python-dotenv
google-generativeai
tqdm
